llm-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.13.3
pydantic==2.6.1
gunicorn==21.2.0
requests==2.31.0
